server/requirements.txt
Flask
flask-cors
flask-pymongo
python-dotenv
PyJWT
bcrypt
cloudinary
gunicorn
